+++ src/main/java/de/maxhenkel/easyvillagers/ClientConfig.java
@@ -8,6 +8,8 @@ public class ClientConfig extends ConfigBase {
     public final ForgeConfigSpec.BooleanValue enableRightClickPickup;
     public final ForgeConfigSpec.DoubleValue villagerVolume;
     public final ForgeConfigSpec.EnumValue<CycleTradesButtonLocation> cycleTradesButtonLocation;
+    public final ForgeConfigSpec.BooleanValue renderBlockContents;
+    public final ForgeConfigSpec.IntValue blockRenderDistance;
 
     public ClientConfig(ForgeConfigSpec.Builder builder) {
         super(builder);
@@ -18,10 +20,16 @@ public class ClientConfig extends ConfigBase {
         villagerVolume = builder
                 .comment("The volume of every villager related sound in this mod")
                 .defineInRange("villager.volume", 1D, 0D, 1D);
-
         cycleTradesButtonLocation = builder
                 .comment("The location of the cycle trades button")
                 .defineEnum("villager.cycle_trades_button_location", CycleTradesButtonLocation.TOP_LEFT);
+        renderBlockContents = builder
+                .comment("Disables rendering any contents of the villager blocks when set to false", "Set to false if you have performance issues with the mod")
+                .worldRestart()
+                .define("villager.render_block_contents", true);
+        blockRenderDistance = builder
+                .comment("The distance in blocks at which the block contents are rendered", "This has no effect if 'render_block_contents' is disabled")
+                .defineInRange("villager.block_render_distance", 32, 1, 256);
     }
 
     public enum CycleTradesButtonLocation {